Successfully created splash screen linked to graph test
I'm planning on playing with a graphing library this weekend to see how
easy it is to create a little pie chart API for us to use to generate
graphs. I created a splash screen that (after 3 seconds) forwards you to
this testing class. Hopefully on Monday, we will start building the main
menu and everything that comes after.