See description, moved test code to valid place
I took my test code out of the SplashActivity and am currently using it
to retrieve the current user and save it into the applicationdata class.
The only thing we need to make sure of, is that every time we make a
change to the user, it gets updated in the global class and in the
database.